changing the ids to long

+++ src/main/java/com/example/egy_tour/repository/ChatbotHistoryRepository.java
@@ -3,6 +3,6 @@ package com.example.egy_tour.repository;
 import com.example.egy_tour.model.ChatbotHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatbotHistoryRepository extends JpaRepository<ChatbotHistory, Integer> {
+public interface ChatbotHistoryRepository extends JpaRepository<ChatbotHistory, Long> {
 
 }
