+++ src/main/java/com/example/egy_tour/repository/LanguageRepository.java
@@ -3,6 +3,6 @@ package com.example.egy_tour.repository;
 import com.example.egy_tour.model.Language;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LanguageRepository extends JpaRepository<Language, Integer> {
+public interface LanguageRepository extends JpaRepository<Language, Long> {
 
 }
