+++ src/main/java/com/example/egy_tour/repository/PreferenceRepository.java
@@ -3,5 +3,5 @@ package com.example.egy_tour.repository;
 import com.example.egy_tour.model.Preference;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PreferenceRepository extends JpaRepository<Preference, Integer> {
+public interface PreferenceRepository extends JpaRepository<Preference, Long> {
 }
