+++ src/main/java/com/example/egy_tour/repository/ProgramRepository.java
@@ -4,6 +4,6 @@ import com.example.egy_tour.model.Program;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ProgramRepository extends JpaRepository<Program, Integer> {
+public interface ProgramRepository extends JpaRepository<Program, Long> {
 
 }
