+++ src/main/java/com/example/egy_tour/repository/TimeSlotRepository.java
@@ -3,5 +3,5 @@ package com.example.egy_tour.repository;
 import com.example.egy_tour.model.TimeSlot;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TimeSlotRepository extends JpaRepository<TimeSlot, Integer> {
+public interface TimeSlotRepository extends JpaRepository<TimeSlot, Long> {
 }
