+++ src/main/java/com/example/egy_tour/repository/TourGuideRepository.java
@@ -4,5 +4,5 @@ package com.example.egy_tour.repository;
 import com.example.egy_tour.model.TourGuide;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TourGuideRepository extends JpaRepository<TourGuide, Integer> {
+public interface TourGuideRepository extends JpaRepository<TourGuide, Long> {
 }
