+++ src/main/java/com/example/egy_tour/repository/TourGuideReviewRepository.java
@@ -3,5 +3,5 @@ package com.example.egy_tour.repository;
 import com.example.egy_tour.model.TourGuideReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TourGuideReviewRepository extends JpaRepository<TourGuideReview, Integer> {
+public interface TourGuideReviewRepository extends JpaRepository<TourGuideReview, Long> {
 }
