+++ src/main/java/com/example/egy_tour/repository/TourismSpotRepository.java
@@ -3,6 +3,6 @@ package com.example.egy_tour.repository;
 import com.example.egy_tour.model.TourismSpot;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TourismSpotRepository extends JpaRepository<TourismSpot, Integer> {
+public interface TourismSpotRepository extends JpaRepository<TourismSpot, Long> {
 
 }
