+++ src/main/java/com/example/egy_tour/repository/TourismSpotReviewRepository.java
@@ -3,5 +3,5 @@ package com.example.egy_tour.repository;
 import com.example.egy_tour.model.TourismSpotReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TourismSpotReviewRepository extends JpaRepository<TourismSpotReview, Integer> {
+public interface TourismSpotReviewRepository extends JpaRepository<TourismSpotReview, Long> {
 }
